Fix status and action when plan succedded
