+++ packages/forklift-console-plugin/locales/en/plugin__forklift-console-plugin.json
@@ -302,6 +302,7 @@
   "Plan name": "Plan name",
   "Plan running": "Plan running",
   "Plane not ready": "Plane not ready",
+  "Plane Succeeded": "Plane Succeeded",
   "Plans": "Plans",
   "Plans for virtualization": "Plans for virtualization",
   "Please choose a NetworkAttachmentDefinition for data transfer.": "Please choose a NetworkAttachmentDefinition for data transfer.",
