+++ packages/forklift-console-plugin/src/modules/Plans/views/create/PlanCreatePage.tsx
@@ -5,75 +5,17 @@ import ProvidersCreateVmMigrationPage from 'src/modules/Providers/views/migrate/
 import { startCreate } from 'src/modules/Providers/views/migrate/reducer/actions';
 import { useFetchEffects } from 'src/modules/Providers/views/migrate/useFetchEffects';
 import { useSaveEffect } from 'src/modules/Providers/views/migrate/useSaveEffect';
-import { ForkliftTrans, useForkliftTranslation } from 'src/utils/i18n';
+import { ForkliftTrans } from 'src/utils/i18n';
 
 import { ProviderModelGroupVersionKind, V1beta1Provider } from '@kubev2v/types';
 import { useK8sWatchResource } from '@openshift-console/dynamic-plugin-sdk';
 import { Alert, PageSection, Title, Wizard } from '@patternfly/react-core';
 
-import { findProviderByID, PlanCreateForm, ProviderVirtualMachinesList } from './components';
-import {
-  PlanCreatePageActionTypes,
-  planCreatePageInitialState,
-  planCreatePageReducer,
-  PlanCreatePageState,
-} from './states';
+import { findProviderByID } from './components';
+import { planCreatePageInitialState, planCreatePageReducer } from './states';
+import { SelectSourceProvider } from './steps';
 
 import './PlanCreatePage.style.css';
-
-export const PlanCreatePage_: React.FC<{
-  namespace: string;
-  filterState: PlanCreatePageState;
-  filterDispatch: React.Dispatch<PlanCreatePageActionTypes>;
-  providers: V1beta1Provider[];
-  selectedProvider: V1beta1Provider;
-}> = ({ filterState, filterDispatch, providers, selectedProvider }) => {
-  const { t } = useForkliftTranslation();
-
-  // Get the ready providers (note: currently forklift does not allow filter be status.phase)
-  const readyProviders = providers.filter((p) => p?.status?.phase === 'Ready');
-
-  const filteredProviders = readyProviders.filter(
-    (provider) =>
-      provider.metadata.name.toLowerCase().includes(filterState.nameFilter.toLowerCase()) &&
-      (filterState.typeFilters.length === 0 ||
-        filterState.typeFilters.includes(provider.spec.type)),
-  );
-
-  const selectedProviderName = selectedProvider?.metadata?.name;
-  const selectedProviderNamespace = selectedProvider?.metadata?.namespace;
-
-  return (
-    <>
-      <Title headingLevel="h2">{t('Select source provider')}</Title>
-
-      <PlanCreateForm
-        providers={filteredProviders}
-        filterState={filterState}
-        filterDispatch={filterDispatch}
-      />
-
-      {filterState.selectedProviderUID && (
-        <>
-          <Title headingLevel="h2" className="forklift--create-plan--title">
-            {t('Select virtual machines')}
-          </Title>
-
-          <ProviderVirtualMachinesList
-            title=""
-            name={selectedProviderName}
-            namespace={selectedProviderNamespace}
-            onSelect={(selectedVms) =>
-              filterDispatch({ type: 'UPDATE_SELECTED_VMS', payload: selectedVms })
-            }
-            initialSelectedIds={filterState.selectedVMs.map((vm) => vm.vm.id)}
-            showActions={false}
-          />
-        </>
-      )}
-    </>
-  );
-};
 
 export const PlanCreatePage: React.FC<{ namespace: string }> = ({ namespace }) => {
   // Get optional initial state context
@@ -111,7 +53,7 @@ export const PlanCreatePage: React.FC<{ namespace: string }> = ({ namespace }) =
       id: 'step-1',
       name: 'Select source provider',
       component: (
-        <PlanCreatePage_
+        <SelectSourceProvider
           namespace={namespace}
           filterState={filterState}
           filterDispatch={filterDispatch}
