+++ packages/forklift-console-plugin/src/modules/Plans/views/create/steps/CreateMigrationPlan/index.ts
@@ -1,0 +1,3 @@
+// @index(['./*', /style/g], f => `export * from '${f.path}';`)
+
+// @endindex
